+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Dropped support for Postgres < 14
 - Dropped support for pg_query < 6
 - Dropped support for Sprockets < 4
+- Dropped support for google-cloud-monitoring < 1
 - Dropped support for system stats for Azure Database (Ruby SDK no longer supported)
 - Dropped support for Linux package for Debian 11